Moved tests for methods not handled by middleware to their own file
